UserService

+++ backend/src/main/kotlin/pwr/diplomaproject/model/dto/UserRolesDto.kt
@@ -1,6 +1,6 @@
 package pwr.diplomaproject.model.dto
 
 data class UserRolesDto(
-    val studentIds: List<Int>,
+    val studentIds: List<Long>,
     val employeeRoles: List<UserEmployeeRoleDto>
 )
